+++ src/tests/e2e/test.ts
@@ -2,20 +2,20 @@ import chalk from "chalk"
 import { exitWithError } from "./helpers/exit-with-error"
 import { setGitNameAndEmail } from "./helpers/set-git-name-and-email"
 import { logTestInfo } from "./test-logging"
-import { testNoFlags } from "./tests/non-interactive/test-no-flags"
-import { testCssModulesWithSassAllFlags } from "./tests/non-interactive/css-modules-with-sass/css-modules-with-sass-all-flags"
-import { testCssModulesWithSassOnly } from "./tests/non-interactive/css-modules-with-sass/css-modules-with-sass-only"
-import { testCssModulesAllFlags } from "./tests/non-interactive/css-modules/css-modules-all-flags"
-import { testCssModulesOnly } from "./tests/non-interactive/css-modules/css-modules-only"
-import { testEmotionAllFlags } from "./tests/non-interactive/emotion/emotion-all-flags"
-import { testEmotionOnly } from "./tests/non-interactive/emotion/emotion-only"
-import { testStyledComponentsAllFlags } from "./tests/non-interactive/styled-components/styled-components-all-flags"
-import { testStyledComponentsOnly } from "./tests/non-interactive/styled-components/styled-components-only"
-import { testTailwindCssAllFlags } from "./tests/non-interactive/tailwind-css/tailwind-css-all-flags"
-import { testTailwindCssOnly } from "./tests/non-interactive/tailwind-css/tailwind-css-only"
-import { testHelpFlag } from "./tests/non-interactive/test-help-flag"
-import { testInvalidInputs } from "./tests/non-interactive/test-invalid-inputs"
-import { testVersionFlag } from "./tests/non-interactive/test-version-flag"
+import { testNoFlags } from "./tests/test-no-flags"
+import { testCssModulesWithSassAllFlags } from "./tests/css-modules-with-sass/css-modules-with-sass-all-flags"
+import { testCssModulesWithSassOnly } from "./tests/css-modules-with-sass/css-modules-with-sass-only"
+import { testCssModulesAllFlags } from "./tests/css-modules/css-modules-all-flags"
+import { testCssModulesOnly } from "./tests/css-modules/css-modules-only"
+import { testEmotionAllFlags } from "./tests/emotion/emotion-all-flags"
+import { testEmotionOnly } from "./tests/emotion/emotion-only"
+import { testStyledComponentsAllFlags } from "./tests/styled-components/styled-components-all-flags"
+import { testStyledComponentsOnly } from "./tests/styled-components/styled-components-only"
+import { testTailwindCssAllFlags } from "./tests/tailwind-css/tailwind-css-all-flags"
+import { testTailwindCssOnly } from "./tests/tailwind-css/tailwind-css-only"
+import { testHelpFlag } from "./tests/test-help-flag"
+import { testInvalidInputs } from "./tests/test-invalid-inputs"
+import { testVersionFlag } from "./tests/test-version-flag"
 ;(async () => {
   // TODO: Find a way to run tests in parallel. Currently failing because simultaneous calls to `npm i -g yarn` or `npm install -g mrm@^3.0.0 mrm-task-lint-staged@^6.0.0` cause crashes.
