+++ src/tests/e2e/tests/css-modules-with-sass/css-modules-with-sass-all-flags.ts
@@ -1,11 +1,11 @@
-import { prettyCommand } from "../../../../../main/helpers/pretty-command"
-import { runCommand } from "../../../../../main/run-command"
-import { checkFormattingLintingBuild } from "../../../helpers/check-formatting-linting-build"
-import { minutesToMilliseconds } from "../../../helpers/minutes-to-milliseconds"
-import { prepareE2eTest } from "../../../helpers/prepare-e2e-test"
-import { logTestInfo } from "../../../test-logging"
+import { prettyCommand } from "../../../../main/helpers/pretty-command"
+import { runCommand } from "../../../../main/run-command"
+import { checkFormattingLintingBuild } from "../../helpers/check-formatting-linting-build"
+import { minutesToMilliseconds } from "../../helpers/minutes-to-milliseconds"
+import { prepareE2eTest } from "../../helpers/prepare-e2e-test"
+import { logTestInfo } from "../../test-logging"
 
-export const testStyledComponentsAllFlags = async (
+export const testCssModulesWithSassAllFlags = async (
   createNextStackDir: string
 ): Promise<void> => {
   const { pathToProdCLI, runDirectory } = await prepareE2eTest(
@@ -16,7 +16,7 @@ export const testStyledComponentsAllFlags = async (
     "--debug",
     "--package-manager=npm",
     "--prettier",
-    "--styling=styled-components",
+    "--styling=css-modules-with-sass",
     "--material-ui",
     "--react-hook-form",
     "--formik",
