+++ src/tests/e2e/tests/tailwind-css/tailwind-css-all-flags.ts
@@ -1,11 +1,11 @@
-import { prettyCommand } from "../../../../../main/helpers/pretty-command"
-import { runCommand } from "../../../../../main/run-command"
-import { checkFormattingLintingBuild } from "../../../helpers/check-formatting-linting-build"
-import { minutesToMilliseconds } from "../../../helpers/minutes-to-milliseconds"
-import { prepareE2eTest } from "../../../helpers/prepare-e2e-test"
-import { logTestInfo } from "../../../test-logging"
+import { prettyCommand } from "../../../../main/helpers/pretty-command"
+import { runCommand } from "../../../../main/run-command"
+import { checkFormattingLintingBuild } from "../../helpers/check-formatting-linting-build"
+import { minutesToMilliseconds } from "../../helpers/minutes-to-milliseconds"
+import { prepareE2eTest } from "../../helpers/prepare-e2e-test"
+import { logTestInfo } from "../../test-logging"
 
-export const testEmotionAllFlags = async (
+export const testTailwindCssAllFlags = async (
   createNextStackDir: string
 ): Promise<void> => {
   const { pathToProdCLI, runDirectory } = await prepareE2eTest(
@@ -16,13 +16,12 @@ export const testEmotionAllFlags = async (
     "--debug",
     "--package-manager=npm",
     "--prettier",
-    "--styling=emotion",
+    "--styling=tailwind-css",
     "--material-ui",
     "--react-hook-form",
     "--formik",
     "--framer-motion",
     "--formatting-pre-commit-hook",
-    "--chakra",
     ".",
   ]
