+++ src/tests/e2e/tests/test-version-flag.ts
@@ -1,0 +1,24 @@
+import { prettyCommand } from "../../../main/helpers/pretty-command"
+import { runCommand } from "../../../main/run-command"
+import { minutesToMilliseconds } from "../helpers/minutes-to-milliseconds"
+import { prepareE2eTest } from "../helpers/prepare-e2e-test"
+import { logTestInfo } from "../test-logging"
+
+export const testVersionFlag = async (
+  createNextStackDir: string
+): Promise<void> => {
+  const { pathToProdCLI, runDirectory } = await prepareE2eTest(
+    createNextStackDir
+  )
+
+  const args = ["--version"]
+
+  logTestInfo("Running command:", prettyCommand(pathToProdCLI, args))
+
+  await runCommand(pathToProdCLI, args, {
+    timeout: minutesToMilliseconds(1),
+    cwd: runDirectory,
+    stdout: "inherit",
+    stderr: "inherit",
+  })
+}
